requirements.txt
# Core AI Dependencies
openai>=1.0.0
anthropic>=0.25.0

# Environment and Configuration
python-dotenv>=1.0.0

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Web Scraping (optional)
requests>=2.28.0
beautifulsoup4>=4.11.0

# OCR Support (optional)
# pytesseract>=0.3.10
# Pillow>=9.5.0

# GUI Support (legacy - optional)
# PyQt6>=6.5.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
